+++ kubejs/server_scripts/Mods/Chipped.js
@@ -1,3 +1,0 @@
-ServerEvents.recipes(event => {
-	event.replaceInput({id: 'chipped:benches/glassblower'}, 'minecraft:blast_furnace', 'minecraft:iron_ingot')
-})