+++ kubejs/startup_scripts/ItemRemoval.js
@@ -205,3 +205,9 @@ removeNamespaceItems('gtceu', [
 	'zpm_cutter',
 	'uv_cutter'
 ])
+
+removeNamespaceItems('minecraft', [
+	'furnace',
+	'blast_furnace',
+	'smoker'
+])